requirements.txt
# Core dependencies
bleak==0.21.1
asyncio-mqtt==0.16.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Machine Learning
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2
plotly==5.15.0
dash==2.11.1
folium==0.14.0

# GUI (optional)
PyQt5==5.15.9

# Web Framework
fastapi==0.100.0
uvicorn[standard]==0.23.1
websockets==11.0.3
pydantic==2.0.3
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.6
sqlalchemy==2.0.19
alembic==1.11.1
asyncpg==0.28.0

# Cache
redis==4.6.0
aioredis==2.0.1

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Logging and Monitoring
loguru==0.7.0
prometheus-client==0.17.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0
mypy==1.4.1
pre-commit==3.3.3

# Utilities
pyyaml==6.0.1
click==8.1.6
rich==13.5.2